working for all angles

+++ material_params.py
@@ -8,7 +8,7 @@ class Quartz(object):
     def __init__(self, frequency_length, run_on_device_decorator):
         self.frequency_length = frequency_length
         self.name = "Quartz"
-        self.frequency = tf.cast(tf.linspace(410.0, 600.0, frequency_length), dtype=tf.complex128)
+        self.frequency = tf.cast(tf.linspace(410.0, 600.0, frequency_length), dtype=tf.complex64)
         self.run_on_device = run_on_device_decorator
 
     @run_on_device
@@ -16,18 +16,18 @@ class Quartz(object):
     
         parameters = {
             "ordinary": {
-                "high_freq": tf.constant(2.356, dtype = tf.complex128),
-                "omega_Tn" : tf.constant([393.5, 450.0, 695.0, 797.0, 1065.0, 1158.0], dtype = tf.complex128),
-                "gamma_Tn" : tf.constant([2.1, 4.5, 13.0, 6.9, 7.2, 9.3], dtype = tf.complex128),
-                "omega_Ln" : tf.constant([403.0, 507.0, 697.6, 810.0, 1226.0, 1155.0], dtype = tf.complex128),
-                "gamma_Ln" : tf.constant([2.8, 3.5, 13.0, 6.9, 12.5, 9.3], dtype = tf.complex128)
+                "high_freq": tf.constant(2.356, dtype = tf.complex64),
+                "omega_Tn" : tf.constant([393.5, 450.0, 695.0, 797.0, 1065.0, 1158.0], dtype = tf.complex64),
+                "gamma_Tn" : tf.constant([2.1, 4.5, 13.0, 6.9, 7.2, 9.3], dtype = tf.complex64),
+                "omega_Ln" : tf.constant([403.0, 507.0, 697.6, 810.0, 1226.0, 1155.0], dtype = tf.complex64),
+                "gamma_Ln" : tf.constant([2.8, 3.5, 13.0, 6.9, 12.5, 9.3], dtype = tf.complex64)
             },
             "extraordinary": {
-                "high_freq": tf.constant(2.383,dtype = tf.complex128),
-                "omega_Tn" : tf.constant([363.5, 487.5, 777.0, 1071.0], dtype = tf.complex128),
-                "gamma_Tn" : tf.constant([4.8, 4.0, 6.7, 6.8], dtype = tf.complex128),
-                "omega_Ln" : tf.constant([386.7, 550.0, 790.0, 1229.0], dtype = tf.complex128),
-                "gamma_Ln" : tf.constant([7.0, 3.2, 6.7, 12.0], dtype = tf.complex128)
+                "high_freq": tf.constant(2.383,dtype = tf.complex64),
+                "omega_Tn" : tf.constant([363.5, 487.5, 777.0, 1071.0], dtype = tf.complex64),
+                "gamma_Tn" : tf.constant([4.8, 4.0, 6.7, 6.8], dtype = tf.complex64),
+                "omega_Ln" : tf.constant([386.7, 550.0, 790.0, 1229.0], dtype = tf.complex64),
+                "gamma_Ln" : tf.constant([7.0, 3.2, 6.7, 12.0], dtype = tf.complex64)
             }
         }
 
@@ -41,7 +41,7 @@ class Quartz(object):
         gamma_Ln_expanded = tf.expand_dims(gamma_Ln, 1)
         omega_Tn_expanded = tf.expand_dims(omega_Tn, 1)
         gamma_Tn_expanded = tf.expand_dims(gamma_Tn, 1)
-        complex_one_j = tf.constant(1j, dtype=tf.complex128)
+        complex_one_j = tf.constant(1j, dtype=tf.complex64)
 
         top_line = omega_Ln_expanded**2. - frequency**2. - complex_one_j * frequency * gamma_Ln_expanded
         bottom_line = omega_Tn_expanded**2. - frequency**2. - complex_one_j * frequency * gamma_Tn_expanded
@@ -89,7 +89,7 @@ class Ambient_Incident_Prism(object):
         
         matrix = tf.stack([element1, element2, element3, element4], axis=1)
 
-        return 0.5 * tf.cast(matrix, dtype=tf.complex128)
+        return 0.5 * tf.cast(matrix, dtype=tf.complex64)
 
 
     @run_on_device
@@ -106,7 +106,7 @@ class Ambient_Incident_Prism(object):
         
         matrix = tf.stack([element1, element2, element3, element4], axis=0)
 
-        return 0.5 * tf.cast(matrix, dtype=tf.complex128)
+        return 0.5 * tf.cast(matrix, dtype=tf.complex64)
 
 
 class Air(object):
@@ -122,5 +122,5 @@ class Air(object):
             [0., 1., 0.],
             [0., 0., 1.]
             ],
-        dtype= tf.complex128)
+        dtype= tf.complex64)
         return tensor
